+++ src/main/java/chess/dao/RoomDao.java
@@ -6,33 +6,146 @@ import chess.dto.response.CurrentTurnDto;
 import chess.dto.response.RoomResponseDto;
 import chess.dto.response.RoomStatusDto;
 import chess.entity.RoomEntity;
+import chess.exception.NotFoundException;
+import java.sql.PreparedStatement;
 import java.util.List;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
 
-public interface RoomDao {
+@Repository
+public class RoomDao {
 
-    RoomEntity findById(int roomId);
+    private final JdbcTemplate jdbcTemplate;
 
-    List<RoomEntity> findAllEntity();
+    public RoomDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
 
-    int save(String roomName, GameStatus gameStatus, Color currentTurn, String password);
+    public RoomEntity findById(final int roomId) {
+        final String sql = "SELECT * FROM room WHERE room_id = ?";
+        final RowMapper<RoomEntity> rowMapper = (resultSet, rowNum) -> new RoomEntity(
+                Integer.parseInt(resultSet.getString("room_id")),
+                resultSet.getString("name"),
+                resultSet.getString("game_status"),
+                resultSet.getString("current_turn"),
+                resultSet.getString("password"),
+                Boolean.parseBoolean(resultSet.getString("is_delete"))
+        );
+        return jdbcTemplate.queryForObject(sql, rowMapper, roomId);
+    }
 
-    boolean isExistName(String roomName);
+    public List<RoomEntity> findAllEntity() {
+        final String sql = "SELECT * FROM room WHERE is_delete = ?";
+        final RowMapper<RoomEntity> rowMapper = (resultSet, rowNum) -> new RoomEntity(
+                Integer.parseInt(resultSet.getString("room_id")),
+                resultSet.getString("name"),
+                resultSet.getString("game_status"),
+                resultSet.getString("current_turn"),
+                resultSet.getString("password"),
+                Boolean.parseBoolean(resultSet.getString("is_delete"))
+        );
+        return jdbcTemplate.query(sql, rowMapper, false);
+    }
 
-    boolean isExistId(int roomId);
+    public int save(final String roomName, final GameStatus gameStatus, final Color currentTurn,
+                    final String password) {
+        final String sql = "INSERT INTO room (name, game_status, current_turn, password) VALUES (?, ?, ?, ?)";
+        final KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(con -> {
+            final PreparedStatement ps = con.prepareStatement(sql, new String[]{"room_id"});
+            ps.setString(1, roomName);
+            ps.setString(2, gameStatus.getValue());
+            ps.setString(3, currentTurn.getValue());
+            ps.setString(4, password);
+            return ps;
+        }, keyHolder);
+        return keyHolder.getKey().intValue();
+    }
 
-    List<RoomResponseDto> findAll();
+    public boolean isExistName(final String roomName) {
+        final String sql = "SELECT EXISTS(SELECT name FROM (SELECT name FROM room WHERE is_delete = ?) AS r WHERE r.name = ?)";
+        return jdbcTemplate.queryForObject(sql, Boolean.class, false, roomName);
+    }
 
-    String findPasswordById(int roomId);
+    public boolean isExistId(final int roomId) {
+        final String sql = "SELECT EXISTS(SELECT room_id FROM room WHERE room_id = ?)";
+        return jdbcTemplate.queryForObject(sql, Boolean.class, roomId);
+    }
 
-    CurrentTurnDto findCurrentTurnById(int roomId);
+    public List<RoomResponseDto> findAll() {
+        final String sql = "SELECT room_id, name, game_status FROM room WHERE is_delete = ?";
+        return jdbcTemplate.query(
+                sql,
+                (resultSet, rowNum) -> RoomResponseDto.of(
+                        Integer.parseInt(resultSet.getString("room_id")),
+                        resultSet.getString("name"),
+                        resultSet.getString("game_status")),
+                false
+        );
+    }
 
-    RoomStatusDto findStatusById(int roomId);
+    public String findPasswordById(final int roomId) {
+        try {
+            final String sql = "SELECT password FROM room WHERE room_id = ? AND is_delete = ?";
+            return jdbcTemplate.queryForObject(sql, String.class, roomId, false);
+        } catch (EmptyResultDataAccessException e) {
+            throw new NotFoundException("방 아이디에 해당하는 비밀번호가 존재하지 않습니다.");
+        }
+    }
 
-    GameStatus findStatus(int roomId);
+    public CurrentTurnDto findCurrentTurnById(final int roomId) {
+        try {
+            final String sql = "SELECT name, current_turn FROM room WHERE room_id = ? AND is_delete = ?";
+            return jdbcTemplate.queryForObject(
+                    sql,
+                    (resultSet, rowNum) -> CurrentTurnDto.of(
+                            resultSet.getString("name"),
+                            Color.from(resultSet.getString("current_turn"))),
+                    roomId, false
+            );
+        } catch (EmptyResultDataAccessException e) {
+            throw new NotFoundException("방 아이디에 해당하는 턴 정보가 존재하지 않습니다.");
+        }
+    }
 
-    int deleteById(int roomId);
+    public RoomStatusDto findStatusById(final int roomId) {
+        try {
+            final String sql = "SELECT name, game_status FROM room WHERE room_id = ? AND is_delete = ?";
+            return jdbcTemplate.queryForObject(
+                    sql,
+                    (resultSet, rowNum) -> RoomStatusDto.of(
+                            resultSet.getString("name"),
+                            resultSet.getString("game_status")),
+                    roomId, false
+            );
+        } catch (EmptyResultDataAccessException e) {
+            throw new NotFoundException("방 아이디에 해당하는 게임 상태가 존재하지 않습니다.");
+        }
+    }
 
-    int updateById(int roomId, GameStatus gameStatus, Color currentTurn);
+    public GameStatus findStatus(final int roomId) {
+        final String sql = "SELECT game_status FROM room WHERE room_id = ?";
+        final RowMapper<GameStatus> rowMapper = (resultSet, rowNum) ->
+                GameStatus.from(resultSet.getString("game_status"));
+        return jdbcTemplate.queryForObject(sql, rowMapper, roomId);
+    }
 
-    int updateStatusById(int roomId, GameStatus gameStatus);
+    public int deleteById(final int roomId) {
+        final String sql = "UPDATE room SET is_delete = ? WHERE room_id = ?";
+        return jdbcTemplate.update(sql, true, roomId);
+    }
+
+    public int updateById(final int roomId, final GameStatus gameStatus, final Color currentTurn) {
+        final String sql = "UPDATE room SET game_status = ?, current_turn = ? WHERE room_id = ?";
+        return jdbcTemplate.update(sql, gameStatus.getValue(), currentTurn.getValue(), roomId);
+    }
+
+    public int updateStatusById(final int roomId, final GameStatus gameStatus) {
+        final String sql = "UPDATE room SET game_status = ? WHERE room_id = ?";
+        return jdbcTemplate.update(sql, gameStatus.getValue(), roomId);
+    }
 }
